requirements.txt
streamlit==1.31.0
ollama==0.1.8
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pymupdf==1.23.21
python-dotenv==1.0.0
validators==0.22.0
nltk==3.8.1
sumy==0.11.0
lxml==4.9.3
html5lib==1.1
Pillow==10.1.0
